Added choice between adaptive and deterministic routing for Bluegene Added choice for receiving multiple messages simulataneously (Bluegene has 6 reception fifos operating simulataneously) Added input speedup variable which constrains the number of input vcs which can be simulataneously transfer data put switch/channel/nic code into seperate files Added util.* files in machine directories so that they can be extended as needed For now Bluegene and IB are working. Simple changes to be made to HypercubeArch and Redstorm too.

+++ examples/pose/HiSim/Routing/DirectionOrdered3D.C
@@ -17,7 +17,7 @@ int DirectionOrdered3D::expectedTime(int s,int d,int ovt,int origovt,int len,int
         Position src,pos;  src.init(s);pos.init(d);
         int extra,expected;
         *hops = abs(pos.x-src.x)+abs(pos.y-src.y)+abs(pos.z-src.y);
-        expected = *hops * config.switchC_Delay + len/config.switchC_BW + START_LATENCY;
+        expected = *hops * config.switchC_Delay + (int)(len/config.switchC_BW) + START_LATENCY;
         extra = (ovt-origovt) - expected;
         if(extra < 0) extra = 0;
         return extra;
